update requestIdleCallback function
