+++ packages/rum/src/domain/record/startFullSnapshots.ts
@@ -1,6 +1,7 @@
 import { LifeCycleEventType, getScrollX, getScrollY, getViewportDimension } from '@datadog/browser-rum-core'
 import type { RumConfiguration, LifeCycle } from '@datadog/browser-rum-core'
 import { timeStampNow, isExperimentalFeatureEnabled, ExperimentalFeature } from '@datadog/browser-core'
+import { requestIdleCallback } from '../../browser/requestIdleCallback'
 import type { BrowserRecord } from '../../types'
 import { RecordType } from '../../types'
 import type { ElementsScrollPositions } from './elementsScrollPositions'
@@ -65,15 +66,10 @@ export function startFullSnapshots(
     return records
   }
   fullSnapshotCallback(takeFullSnapshot())
-  let requestIdleCallbackId: number | undefined
   const { unsubscribe } = lifeCycle.subscribe(LifeCycleEventType.VIEW_CREATED, (view) => {
     flushMutations()
-
-    if (requestIdleCallbackId !== undefined) {
-      cancelIdleCallback(requestIdleCallbackId)
-    }
     if (isExperimentalFeatureEnabled(ExperimentalFeature.ASYNC_FULL_SNAPSHOT)) {
-      requestIdleCallbackId = requestIdleCallback(() => {
+      requestIdleCallback(() => {
         fullSnapshotCallback(
           takeFullSnapshot(view.startClocks.timeStamp, {
             shadowRootsController,
